+++ plugin-maven/src/main/java/com/diffplug/spotless/maven/kotlin/Ktfmt.java
@@ -44,11 +44,14 @@ public class Ktfmt implements FormatterStepFactory {
 	@Parameter
 	private Boolean removeUnusedImports;
 
+	@Parameter
+	private Boolean manageTrailingCommas;
+
 	@Override
 	public FormatterStep newFormatterStep(FormatterStepConfig config) {
 		String version = this.version != null ? this.version : KtfmtStep.defaultVersion();
 		Style style = this.style != null ? Style.valueOf(this.style) : null;
-		KtfmtFormattingOptions options = new KtfmtFormattingOptions(maxWidth, blockIndent, continuationIndent, removeUnusedImports);
+		KtfmtFormattingOptions options = new KtfmtFormattingOptions(maxWidth, blockIndent, continuationIndent, removeUnusedImports, manageTrailingCommas);
 		return KtfmtStep.create(version, config.getProvisioner(), style, options);
 	}
 }